Homework 2 bby
This is my homework 2, where we recreated the card app from class. This
assignment has 10 cards with images, titles, descriptions and three
buttons at the bottom that are share, favorite and delete. The delete
button deletes the card you click on. There is an array at the top of
each card that dynamically counts the current cards.